pyproject.toml
[project]
name = "habitantes-grenoble-agent"
version = "0.1.0"
description = "FastAPI service for Brazilian expats chatbot in Grenoble"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "langgraph>=0.0.26",
    "langchain-openai>=0.0.5",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "qdrant-client>=1.7.0",
    "sentence-transformers>=2.3.0",
    "loguru>=0.7.2",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.1",
    "fastembed>=0.2.0",
    "python-telegram-bot>=21.0",
    "httpx>=0.27.0",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.1.0",
    "isort>=5.13.0",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
]

[tool.setuptools.packages.find]
where = ["api/src"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100
